+++ tests/integrationtests/api/test_ohsome_timeout.py
@@ -2,10 +2,9 @@
 Testing FastAPI Applications:
 https://fastapi.tiangolo.com/tutorial/testing/
 
-Shared tests for `/indicator` and `/report` endpoints using the `bpolys` parameter.
+Shared tests for `/indicator` endpoints using the `bpolys` parameter.
 Tests for the individual endpoints and using the `bpolys` parameter please see:
     - `test_api_indicator_geojson_io.py`
-    - `test_api_report_geojson_io.py`
 """
 
 import os
@@ -35,22 +34,12 @@ def test_ohsome_timeout(client, bpolys):
             request=httpx.Request("POST", "https://www.example.org/"),
         )
 
-        for endpoint, parameters in (
-            (
-                "/reports/minimal",
-                {
-                    "bpolys": bpolys,
-                },
-            ),
-            (
-                "/indicators/minimal",
-                {
-                    "bpolys": bpolys,
-                    "topic": "minimal",
-                },
-            ),
-        ):
-            response = client.post(endpoint, json=parameters)
-            assert response.status_code == 422
-            content = response.json()
-            assert content["type"] == "OhsomeApiError"
+        endpoint = "/indicators/minimal"
+        parameters = {
+            "bpolys": bpolys,
+            "topic": "minimal",
+        }
+        response = client.post(endpoint, json=parameters)
+        assert response.status_code == 422
+        content = response.json()
+        assert content["type"] == "OhsomeApiError"