+++ tests/unittests/test_definitions.py
@@ -6,7 +6,6 @@ from ohsome_quality_api import definitions
 from ohsome_quality_api.indicators.models import (
     IndicatorMetadata as IndicatorMetadata,
 )
-from ohsome_quality_api.reports.base import ReportMetadata as ReportMetadata
 
 
 class TestDefinitions(unittest.TestCase):
@@ -34,13 +33,6 @@ def test_load_metadata_indicator():
         assert isinstance(v, IndicatorMetadata)
 
 
-def test_load_metadata_report():
-    metadata = definitions.load_metadata("reports")
-    assert isinstance(metadata, dict)
-    for v in metadata.values():
-        assert isinstance(v, ReportMetadata)
-
-
 def test_load_metadata_wrong_module():
     with pytest.raises(AssertionError):
         definitions.load_metadata("foo")
@@ -51,13 +43,6 @@ def test_load_metadata_wrong_module():
 def test_get_metadata_indicator():
     metadata = definitions.get_metadata("indicators", "Minimal")
     assert isinstance(metadata, IndicatorMetadata)
-
-
-def test_get_metadata_report():
-    metadata = definitions.get_metadata("reports", "Minimal")
-    assert isinstance(metadata, ReportMetadata)
-    with pytest.raises(KeyError):
-        definitions.get_metadata("reports", "")
 
 
 def test_get_metadata_wrong_class():